All tests are now passing with the new Twisted trial runner.

+++ src/utils/test_utils.py
@@ -1,8 +1,10 @@
 """
-Assorted utilities for unit testing.
+Assorted utilities for unit testing. Lots of assorted unholy stuff going on
+in here, so not for the faint of heart.
 """
 
 import os
+
 import psycopg2
 from twisted.trial import unittest
 from twisted.internet.defer import inlineCallbacks
@@ -13,6 +15,11 @@ from src.daemons.server.commands.handler import CommandHandler
 from src.daemons.server.objects.object_store import ObjectStore
 from src.accounts.account_store import AccountStore
 from src.daemons.proxy.sessions.session_manager import SessionManager
+
+# I'm probably going to hell for this, but we use it to make sure that
+# we only create the test DB once, instead of between every test. Makes things
+# run a little faster.
+DB_WAS_CREATED = False
 
 
 #noinspection PyDocstring,PyPep8Naming
@@ -42,11 +49,25 @@ class MockMudService(object):
 
     @inlineCallbacks
     def prep_and_load(self):
+        """
+        Gets the various stores and other stuff set up and loaded from the DB.
+        """
+
         yield self.object_store.prep_and_load()
         yield self.account_store.prep_and_load()
 
     #noinspection PyProtectedMember
+    @inlineCallbacks
     def unload(self):
+        """
+        Right now this is specific to unit tests, and doesn't actually exist
+        in the actual MudService class. We clear out the various test tables
+        and close all DB pointers.
+        """
+
+        yield self.object_store.db_manager._db.runOperation(
+            "TRUNCATE dott_accounts, dott_objects")
+
         self.object_store.db_manager._db.close()
         self.account_store.db_manager._db.close()
 
@@ -61,16 +82,23 @@ class DottTestCase(unittest.TestCase):
     def setUp(self):
         """
         By default, create a bare minimal set of data stores.
+
+        .. note:: If you override this, you'll need to make sure that your
+            new method returns a deferred. yield super() if inlinecallbacks.
         """
 
         yield self.create_clean_game_env()
 
+    @inlineCallbacks
     def tearDown(self):
         """
-        Delete the created data stores between unit tests.
+        Cleans up the created environment.
+
+        .. note:: If you override this, you'll need to make sure that your
+            new method returns a deferred. yield super() if inlinecallbacks.
         """
 
-        self.cleanup_game_env()
+        yield self.mud_service.unload()
 
     @inlineCallbacks
     def create_clean_game_env(self):
@@ -78,25 +106,33 @@ class DottTestCase(unittest.TestCase):
         Creates a fresh set of stores, DBs, and etc.
         """
 
-        conn = psycopg2.connect(user=settings.DATABASE_USERNAME)
-        conn.set_session(autocommit=True)
-        cur = conn.cursor()
-        cur.execute("DROP DATABASE IF EXISTS %s;" % settings.TEST_DATABASE_NAME)
-        cur.execute("CREATE DATABASE %s WITH OWNER=%s;" % (
-            settings.TEST_DATABASE_NAME, settings.TEST_DATABASE_USERNAME
-        ))
-        cur.close()
-        conn.close()
+        global DB_WAS_CREATED
+        if not DB_WAS_CREATED:
+            # We have to use psycopg2 directly since txpostgres can't
+            # be used with autocommit=True.
+            conn = psycopg2.connect(user=settings.DATABASE_USERNAME)
+            conn.set_session(autocommit=True)
+            cur = conn.cursor()
+            # Destroy and re-create the test DB to make sure the
+            # schema is current.
+            cur.execute("DROP DATABASE IF EXISTS %s;" % settings.TEST_DATABASE_NAME)
+            cur.execute("CREATE DATABASE %s WITH OWNER=%s;" % (
+                settings.TEST_DATABASE_NAME, settings.TEST_DATABASE_USERNAME
+            ))
+            cur.close()
+            conn.close()
 
-        conn = psycopg2.connect(
-            user=settings.DATABASE_USERNAME, database=settings.TEST_DATABASE_NAME)
-        conn.set_session(autocommit=True)
-        cur = conn.cursor()
-        schema_file = os.path.join(settings.BASE_PATH, 'misc', 'dott-schema.sql')
-        schema = open(schema_file).read()
-        cur.execute(schema)
-        cur.close()
-        conn.close()
+            conn = psycopg2.connect(
+                user=settings.DATABASE_USERNAME, database=settings.TEST_DATABASE_NAME)
+            conn.set_session(autocommit=True)
+            cur = conn.cursor()
+            # Load the schema from the export.
+            schema_file = os.path.join(settings.BASE_PATH, 'misc', 'dott-schema.sql')
+            schema = open(schema_file).read()
+            cur.execute(schema)
+            cur.close()
+            conn.close()
+            DB_WAS_CREATED = True
 
         self.mud_service = MockMudService()
         yield self.mud_service.prep_and_load()
@@ -105,10 +141,3 @@ class DottTestCase(unittest.TestCase):
         self.session_manager = self.mud_service.session_manager
         self.object_store = self.mud_service.object_store
         self.account_store = self.mud_service.account_store
-
-    def cleanup_game_env(self):
-        """
-        Cleans up the created environment.
-        """
-
-        self.mud_service.unload()
